requirements.txt
Flask>=2.2
Flask-Login
Flask-Migrate
Flask-WTF
python-dotenv
mysql-connector-python
SQLAlchemy
alembic
Werkzeug
Flask-SQLAlchemy


